backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
aiohttp==3.9.1
alembic==1.13.1
python-multipart==0.0.6
